pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "serial-tool"
version = "0.1.0"
description = "XCOM V2.6 compatible serial debugger with CLI enhancements"
authors = [{name = "OpenCom Team", email = "team@opencom.dev"}]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["serial", "debugger", "uart", "xcom", "cli", "web"]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pyserial>=3.5",
    "python-multipart>=0.0.6",
    "click>=8.1.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "black>=23.0", "isort>=5.0", "mypy>=1.0"]

[project.entry-points."console_scripts"]
serial-tool = "backend.cli.main:cli"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*", "shared*"]

[tool.black]
line-length = 88
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]

[tool.mypy]
files = ["backend", "shared"]
ignore_missing_imports = true
show_error_codes = true
